app/requirements.txt
fastapi==0.105.0
sqlalchemy==2.0.3
pydantic==2.8.2
passlib==1.7.4
alembic==1.9.2
bcrypt==3.2.0
PyJWT==2.9.0
psycopg2-binary==2.9.5
requests==2.32.3
python-dotenv==1.0.1
pyjwt[crypto]==2.9.0
python-multipart==0.0.9
pandas==1.4.3
numpy==1.26.4
asyncpg==0.29.0
uvicorn==0.20.0
SQLAlchemy-Utils==0.41.2
